# Label: Changed regexp.go file and added a single test case 'ſſs'