Fixes #14066

Added a test case "ſſs" and changed Regexp.go file , such that all the test case will pass.