@colega 2024-05-13
As [I mentioned in my comment](https://github.com/prometheus/prometheus/issues/14066#issuecomment-2100048592) this looks like the least performant way to implement this, and I'd guess that it can be even slower than running the original regular expression before optimization.

I would run some benchmarks to see the performance impact of looping here.