@beorn7 2024-06-04
Note that there is now https://github.com/prometheus/prometheus/pull/14170